+++ linqpad/Export_BloggerSearchIndex_FromPosts.linq
@@ -10,9 +10,9 @@ bool DEBUG_MODE = false;
 // INPUT OUTPUT SETTINGS
 string BLOGGER_XML_DIRECTORY = @"C:\Users\KAINENG\Downloads\";
 string ARCHIVE_XML_DIRECTORY = @"C:\Users\KAINENG\Documents\LINQPad Queries\blog-archive\";
-string OUTPUT_DIRECTORY = @"C:\Users\KAINENG\Documents\GitHub\knneo.github.io\blog\";
+string OUTPUT_DIRECTORY = @"C:\Users\KAINENG\Documents\GitHub\knreports\";
 string OUTPUT_DIRECTORY_SUBFOLDER = "posts";
-string HOMEPAGE_FILENAME = @"C:\Users\KAINENG\Documents\GitHub\knneo.github.io\blog\js\searchIndex.js";
+string HOMEPAGE_FILENAME = @"C:\Users\KAINENG\Documents\GitHub\knreports\js\searchIndex.js";
 
 // PROGRAM SETTINGS
 bool HOMEPAGE_ONLY = false;
@@ -25,10 +25,12 @@ XNamespace DEFAULT_XML_NAMESPACE = XNamespace.Get("http://www.w3.org/2005/Atom")
 int MIN_TOKEN_LENGTH = 3;
 List<string> TOKEN_STOP_WORDS = new List<string>() { "the", "a", "is", "of" };
 List<string> TOKEN_SPLIT_CHARACTERS = new List<string> { " ", ".", ",", "!", "?", "&nbsp;", "&quot;", "*", ":", ";", "-", "(", ")", "[", "]", "\"" };
-bool GENERATE_SLUG_BY_POST_TITLE = false;
+bool GENERATE_SLUG_BY_POST_TITLE = true;
+int GENERATE_SLUG_MAX_LENGTH = 70;
+string BLOGGER_XML_RENAME_SUFFIX = "knreports";
 
 // POST SETTINGS
-List<String> POST_IGNORE_TAGS = new List<string>() { "The Archive" };
+List<String> POST_IGNORE_TAGS = new List<string>() { "The Archive", "The Statement" };
 string POSTS_SINCE = "2000-01-01";
 
 void Main()
@@ -48,6 +50,7 @@ void Main()
 	Console.WriteLine("Done.");
 }
 
+
 string[] GetBloggerXmlFilePath(string inputPath, string backupPath)
 {
     Console.WriteLine("Reading Config...");	
@@ -59,13 +62,16 @@ string[] GetBloggerXmlFilePath(string inputPath, string backupPath)
 	{
         if(DEBUG_MODE) Console.WriteLine($"Single xml source found; Moving file to {backupPath}");		
 	    string[] dests = Directory.GetFiles(Path.GetDirectoryName(backupPath), "blog-*.xml");
-	    if(dests.Length == 1)
+	    if(DEBUG_MODE) Console.WriteLine("Destination files found; Moving all files to archive");
+	    foreach(var dest in dests)
 		{
-	        if(DEBUG_MODE) Console.WriteLine("Destination file found; Moving to archive");
-        	File.Delete(dests[0].Replace(backupPath, $"{backupPath}archive\\"));
-        	File.Move(dests[0], dests[0].Replace(backupPath, $"{backupPath}archive\\"));
+			if(dest.Contains(BLOGGER_XML_RENAME_SUFFIX))
+			{
+	        	File.Delete(dest.Replace(backupPath, $"{backupPath}archive\\"));
+	        	File.Move(dest, dest.Replace(backupPath, $"{backupPath}archive\\"));
+			}
 		}
-        File.Move(sources[0], sources[0].Replace(inputPath, backupPath));
+        File.Move(sources[0], sources[0].Replace(inputPath, backupPath).Replace(".xml", "-" + BLOGGER_XML_RENAME_SUFFIX + ".xml"));
 	}
     else if(sources.Length == 0)
     {
@@ -78,12 +84,15 @@ string[] GetBloggerXmlFilePath(string inputPath, string backupPath)
 	    if(DEBUG_MODE) Console.WriteLine("Destination files found; Moving all files to archive");
 	    foreach(var dest in dests)
 		{
-        	File.Delete(dest.Replace(backupPath, $"{backupPath}archive\\"));
-        	File.Move(dest, dest.Replace(backupPath, $"{backupPath}archive\\"));
+			if(dest.Contains(BLOGGER_XML_RENAME_SUFFIX))
+			{
+	        	File.Delete(dest.Replace(backupPath, $"{backupPath}archive\\"));
+	        	File.Move(dest, dest.Replace(backupPath, $"{backupPath}archive\\"));
+			}
 		}
 	    foreach(var source in sources)
 		{
-        	File.Move(source, source.Replace(inputPath, backupPath));
+        	File.Move(source, source.Replace(inputPath, backupPath).Replace(".xml", "-" + BLOGGER_XML_RENAME_SUFFIX + ".xml"));
 		}
     }
 	// Read xml files to process
@@ -99,7 +108,7 @@ string[] GetBloggerXmlFilePath(string inputPath, string backupPath)
     }
     else
     {
-        throw new NotSupportedException("More than 1 xml files found; Appending all files for process");
+        if(DEBUG_MODE) Console.WriteLine("More than 1 xml files found; Appending all files for process");
     }
 	
 	return xmls;
@@ -176,7 +185,7 @@ SearchIndex GenerateSearchIndex(List<XElement> xmlPosts)
 		if(pageTagsXml.Any(xml => POST_IGNORE_TAGS.Contains(xml)))
 			continue;
 		// Create output page link and index in linked list
-        var pageLink = "../" + Path.GetFileNameWithoutExtension(BLOGGER_XML_DIRECTORY.Replace(BLOGGER_XML_DIRECTORY, OUTPUT_DIRECTORY_SUBFOLDER)) + "/" + publishDate.Year.ToString("0000") + "/"  + publishDate.Month.ToString("00") + "/"  + (GENERATE_SLUG_BY_POST_TITLE ? generatedLink : Path.GetFileNameWithoutExtension(bloggerLink)) + "/index." + postExtension;
+        var pageLink = "" + Path.GetFileNameWithoutExtension(BLOGGER_XML_DIRECTORY.Replace(BLOGGER_XML_DIRECTORY, OUTPUT_DIRECTORY_SUBFOLDER)) + "/" + publishDate.Year.ToString("0000") + "/"  + publishDate.Month.ToString("00") + "/"  + (GENERATE_SLUG_BY_POST_TITLE ? generatedLink : Path.GetFileNameWithoutExtension(bloggerLink)) + "/index." + postExtension;
 
 		// Generate index
 		// Remove extra tags, filter by special characters and whitespace, lowercase, other filters (if any)
@@ -234,14 +243,15 @@ void GenerateSearchIndexFile(SearchIndex searchIndex)
     File.WriteAllText(HOMEPAGE_FILENAME, export);
 }
 
-string GenerateSlug(string title, int maxLength = 45)
+string GenerateSlug(string title)
 {
 	string slug = title.ToLower();
+	slug = slug.Replace("&quot;","");
 	slug = Regex.Replace(slug, @"\s+", "-");
 	slug = Regex.Replace(slug, @"[^a-z0-9\-_]", "");
-	slug = Regex.Replace(slug, @"\b\d(?!\d)", "");
-	slug = slug.Replace("--","-").Trim('-');
-	return slug.Length > maxLength ? slug.Substring(0, slug.Substring(0, maxLength).LastIndexOf('-')) : slug;
+	//slug = Regex.Replace(slug, @"\b\d(?!\d)", "");
+	slug = slug.Replace("--","-").Replace("--","-").Trim('-');
+	return slug.Length > GENERATE_SLUG_MAX_LENGTH ? slug.Substring(0, slug.Substring(0, GENERATE_SLUG_MAX_LENGTH).LastIndexOf('-')) : slug;
 }
 
 public class SearchIndex {
